+++ Addon/ElvUI_mMediaTag/core/changelog/3.71.lua
@@ -1,0 +1,7 @@
+mMT.Changelog[371] = {
+	DATE = "18.03.2025",
+	IMPORTANT = { "Attention! With the next version 4.0 of mMT there will be some major changes that will reset some settings." },
+	FIX = {
+		"Bug with unknown and set Favorite Teleports.",
+	},
+}